feat: add storybook and small changes on component
